requirements.txt
pandas
requests
scikit-learn
arxiv
openai
matplotlib
black==25.9.0
pre-commit
tqdm
torch
semanticscholar
Langfuse
python-dotenv
bibtexparser
ollama
groq